refactor: Breakout subcommands opts into different files
