+++ src/cli/clear_kinetics_opts.rs
@@ -1,0 +1,12 @@
+use super::super::utils::InputBam;
+use clap::Args;
+use std::fmt::Debug;
+
+#[derive(Args, Debug)]
+pub struct ClearKineticsOptions {
+    #[clap(flatten)]
+    pub input: InputBam,
+    /// Output bam file without hifi kinetics
+    #[clap(default_value = "-")]
+    pub out: String,
+}
