+++ src/cli/ddda_to_m6a_opts.rs
@@ -1,0 +1,12 @@
+use super::super::utils::InputBam;
+use clap::Args;
+use std::fmt::Debug;
+
+#[derive(Args, Debug)]
+pub struct DddaToM6aOptions {
+    #[clap(flatten)]
+    pub input: InputBam,
+    /// Output bam file
+    #[clap(default_value = "-")]
+    pub out: String,
+}
